refactor(cell-decomp): extract PDK-neutral Verilog parse + AIG decomp into a shared crate
ADR 0019 step C1.1. Lifts the PDK-neutral cell-logic machinery out of
jacquard core into a new `crates/cell-decomp` so both core and the future
Liberty->IR converter (C1.3) can link it. Pure refactor — no behaviour
change (304 lib tests unchanged).

Moved into cell-decomp:
- the Verilog parser AST + parsers (BehavioralModel/Gate, UdpModel/Row,
  parse_functional_model, parse_udp) — relocated from src/sky130_pdk.rs,
  which only happened to host these PDK-neutral types;
- the AIG-build primitives + DecompResult (formerly src/pdk_decomp.rs,
  now deleted);
- the gate-level evaluator oracle (eval_behavioral_model / eval_udp_for_inputs).

Left in core (consumption side): CellInputs, decompose_from_behavioral,
decompose_with_pdk, load_pdk_models, and the per-PDK classifiers.

Cleanups folded in (via /simplify):
- eval_behavioral_model now dispatches UDPs by membership in the model's
  `udps` map instead of a hardcoded `sky130_fd_sc_hd__udp_` prefix, so the
  shared crate is genuinely vendor-agnostic (GF180/IHP UDPs route too) —
  behaviour-identical for SKY130 whose UDPs are keyed in that map;
- GATE_MARKER narrowed to pub(crate) (encoding stays encapsulated);
- stale comments referencing the deleted pdk_decomp module fixed.

Co-developed-by: Claude Code v2.1.187 (claude-opus-4-8)
