+++ crates/cell-decomp/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "cell-decomp"
+version = "0.1.0"
+edition = "2021"
+description = "PDK-neutral behavioral Verilog parsing and AIG decomposition primitives shared by jacquard core and the Liberty->IR converter. See docs/adr/0019-cell-model-ir.md."
+license = "Apache-2.0"
+repository = "https://github.com/gpu-eda/Jacquard"
+publish = false
+
+[lib]
+path = "src/lib.rs"
