+++ src/lib.rs
@@ -24,7 +24,9 @@
 //! - [`aigpdk`] — AIGPDK standard cell library interface (AND gates, DFFs, clock gates, SRAMs)
 //! - [`sky130`] / [`sky130_pdk`] — SKY130 cell library support with Liberty timing
 //! - [`gf180mcu`] / [`gf180mcu_pdk`] — GF180MCU cell library support (7t/9t variants; in progress)
-//! - [`pdk`] / [`pdk_decomp`] — PDK-shared types, library detection, and behavioural-model primitives
+//! - [`pdk`] — PDK-shared types and library detection. The behavioural-model
+//!   parsing and AIG-decomposition primitives live in the standalone
+//!   `cell_decomp` crate (`crates/cell-decomp`).
 //! - [`aig`] — And-inverter graph construction from [`netlistdb::NetlistDB`]
 //! - [`liberty_parser`] — Liberty (.lib) file parser for cell timing data
 //! - [`staging`] — Splits the AIG into pipeline stages via `--level-split` thresholds
@@ -52,8 +54,6 @@ pub mod gf180mcu_pdk;
 pub mod pdk;
 
 pub mod cell_library;
-
-pub mod pdk_decomp;
 
 pub mod aig;
